Refactor getting devices for less code dupe
